backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
SQLAlchemy==2.0.37
psycopg2-binary==2.9.9
ortools==9.15.6755
httpx==0.27.2
